requirements.txt
fastapi
uvicorn
requests
python-multipart
openai-whisper
openai
ffmpeg-python
pydub
